Moved Main Menu from main to Menus.hpp file and made basic pause menu
