+++ Makefile
@@ -1,6 +1,6 @@
 SRC=src
 MAINSRC=$(SRC)/main.cpp
-HEADERS= $(SRC)/Exception.hpp $(SRC)/Game.hpp $(SRC)/MediaManager.hpp $(SRC)/Particle.hpp $(SRC)/Animation.hpp $(SRC)/Wave.hpp $(SRC)/Player.hpp $(SRC)/NPC.hpp $(SRC)/Config.hpp $(SRC)/Character.hpp $(SRC)/Tile.hpp $(SRC)/Map.hpp $(SRC)/Lightning.hpp
+HEADERS= $(SRC)/Exception.hpp $(SRC)/Game.hpp $(SRC)/MediaManager.hpp $(SRC)/Particle.hpp $(SRC)/Animation.hpp $(SRC)/Wave.hpp $(SRC)/Player.hpp $(SRC)/NPC.hpp $(SRC)/Config.hpp $(SRC)/Character.hpp $(SRC)/Tile.hpp $(SRC)/Map.hpp $(SRC)/Lightning.hpp $(SRC)/Menus.hpp
 
 LINUXFLAGS=-I/usr/include/SDL2 -D_REENTRANT
 LINUXLIBS=-lSDL2 -lSDL2_mixer -lSDL2_ttf
